implemented generation and submission of ci job script after user input

+++ py-lib/add.py
@@ -1,8 +1,10 @@
 from subprocess import run
 import os
+import sys
 import re
 from pathlib import Path
-
+from gen_job_script import write_job_script
+    
 # Read input to add from add.in if it exists, otherwise create it
 try:
     open("add.in", 'rb').read()
@@ -35,7 +37,7 @@ if odd_exists:
 # Cleanup - remove add.in
 run("rm add.in", shell=True)
 
-# Create even and odd directories
+# Ask if user wants even and odd directories to be generated
 gen_dir = eval(re.sub('(no|No|n|N|false)', 'False', re.sub('(yes|Yes|y|Y|true)', 'True', str(input("Generate directories? ")))))
 
 if gen_dir:
@@ -43,8 +45,10 @@ if gen_dir:
     for dirs in ['EVEN','ODD']:
         Path(dir_path+'/'+dirs).mkdir(parents=True, exist_ok=True)
         os.chdir(dirs)
-        run("cp ../HFD.DAT .", shell=True)
-        run("cp ../CONF" + dirs.lower() + ".INP CONF.INP", shell=True)
+        if os.path.isfile('../HFD.DAT'):
+            run("cp ../HFD.DAT .", shell=True)
+        if os.path.isfile('../CONF' + dirs.lower() + '.INP'):
+            run("cp ../CONF" + dirs.lower() + ".INP CONF.INP", shell=True)
         if os.path.isfile('../SGC.CON'):
             run("cp ../SGC.CON .", shell=True)
         if os.path.isfile('../SCRC.CON'):
@@ -58,6 +62,36 @@ if gen_dir:
                 f.write('0, 0, 0, 0, 1')
                 f.close()
         os.chdir('../')
+    
+    # Ask if user wants to submit ci jobs
+    run_ci = eval(re.sub('(no|No|n|N|false)', 'False', re.sub('(yes|Yes|y|Y|true)', 'True', str(input("Submit CI jobs? ")))))
+    if run_ci:
+        for dirs in ['EVEN','ODD']:
+            # Change to working directory
+            os.chdir(dirs)
+            
+            # Check if files required for CI exist
+            if not os.path.isfile('HFD.DAT'):
+                print('HFD.DAT is missing from ' + dirs + ' directory')
+                sys.exit()
+            if not os.path.isfile('CONF.INP'):
+                print('CONF.INP is missing from ' + dirs + ' directory')
+                sys.exit()
+            if not os.path.isfile('c.in'):
+                print('c.in is missing from ' + dirs + ' directory')
+                sys.exit()
+                
+            # Generate new job script if it doesn't exist yet
+            if not os.path.isfile('ci.qs'):
+                print('generating new ci.qs in ' + dirs + ' directory')
+                write_job_script('ci', 5, 64, True, 0, 'safrono')
+            
+            # Submit job script
+            run("sbatch ci.qs", shell=True)
+                
+            # Change to root directory
+            os.chdir('../')
+
 
 print('add script completed')
